+++ inc/tp_image_utils_functions/DeNoise.h
@@ -12,12 +12,12 @@ namespace tp_image_utils_functions
 struct ByteRegion
 {
   uint8_t value{0};
-  int count{0};
+  size_t count{0};
 
-  int minX{0}; //!< The min x coordinate of this region calculated by calculateBoundingBoxes()
-  int minY{0}; //!< The min y coordinate of this region calculated by calculateBoundingBoxes()
-  int maxX{0}; //!< The max x coordinate of this region calculated by calculateBoundingBoxes()
-  int maxY{0}; //!< The max y coordinate of this region calculated by calculateBoundingBoxes()
+  size_t minX{0}; //!< The min x coordinate of this region calculated by calculateBoundingBoxes()
+  size_t minY{0}; //!< The min y coordinate of this region calculated by calculateBoundingBoxes()
+  size_t maxX{0}; //!< The max x coordinate of this region calculated by calculateBoundingBoxes()
+  size_t maxY{0}; //!< The max y coordinate of this region calculated by calculateBoundingBoxes()
 };
 
 //##################################################################################################
@@ -29,8 +29,8 @@ struct ByteRegions
   //! The region indexes for each pixel in the image
   std::vector<int> map;
 
-  int w{0};
-  int h{0};
+  size_t w{0};
+  size_t h{0};
 
   //################################################################################################
   //! Separate the regions of an image
@@ -53,7 +53,7 @@ struct ByteRegions
 
 //##################################################################################################
 tp_image_utils::ByteMap deNoise(const tp_image_utils::ByteMap& src,
-                                int minSize,
+                                size_t minSize,
                                 bool addCorners,
                                 uint8_t solid=0,
                                 uint8_t space=255);
@@ -64,17 +64,17 @@ tp_image_utils::ByteMap deNoiseBlobs(const tp_image_utils::ByteMap& src,
                                      float maxAspectRatio,
                                      float minDensity,
                                      float maxDensity,
-                                     int minSize,
-                                     int maxSize,
+                                     size_t minSize,
+                                     size_t maxSize,
                                      bool addCorners,
                                      uint8_t solid=0,
                                      uint8_t space=255);
 
 //##################################################################################################
-tp_image_utils::ByteMap deNoiseStripes(const tp_image_utils::ByteMap& src, int minSize, uint8_t solid=0, uint8_t space=255);
+tp_image_utils::ByteMap deNoiseStripes(const tp_image_utils::ByteMap& src, size_t minSize, uint8_t solid=0, uint8_t space=255);
 
 //##################################################################################################
-tp_image_utils::ByteMap deNoiseKnoblets(const tp_image_utils::ByteMap& src, int knobletWidth, uint8_t solid=0, uint8_t space=255);
+tp_image_utils::ByteMap deNoiseKnoblets(const tp_image_utils::ByteMap& src, size_t knobletWidth, uint8_t solid=0, uint8_t space=255);
 
 
 }
